Feat:테스트 코드 작성 및 페이지네이션 적용

+++ src/main/java/com/sparta/sportify/dto/cash/response/CashLogsResponseDto.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class CashLogsResponseDto {
-	 private int amount;
+	 private Long amount;
 	 private LocalDateTime createAt;
 	 private CashType type;
 }
